Moved the layout components into a Layout folder
